Add-AutoUpdate-in-Chat-Features-ver5
※注意：メッセージ送信時に
AttributeError: '_thread._local' object has no attribute 'value'
が発生します。
