+++ error.txt
@@ -1,0 +1,41 @@
+127.0.0.1 - - [20/Sep/2019 08:54:46] "POST /post_api HTTP/1.1" 200 -
+
+----- Generating word after Epoch:
+----- diversity: 0.2
+----- Generating with seed: "いいね"
+127.0.0.1 - - [20/Sep/2019 08:54:50] "POST /post_ajax HTTP/1.1" 500 -
+Traceback (most recent call last):
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 2463, in __call__
+    return self.wsgi_app(environ, start_response)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 2449, in wsgi_app
+    response = self.handle_exception(e)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 1866, in handle_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 2446, in wsgi_app
+    response = self.full_dispatch_request()
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 1951, in full_dispatch_request
+    rv = self.handle_user_exception(e)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 1820, in handle_user_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 1949, in full_dispatch_request
+    rv = self.dispatch_request()
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/flask/app.py", line 1935, in dispatch_request
+    return self.view_functions[rule.endpoint](**req.view_args)
+  File "/Users/kdr/python_project/flaskblog/run.py", line 166, in post_ajax
+    bot_return = bot_run(content)
+  File "/Users/kdr/python_project/flaskblog/run.py", line 123, in bot_run
+    preds = model.predict(x_pred, verbose=0)[0]
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/keras/engine/training.py", line 1456, in predict
+    self._make_predict_function()
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/keras/engine/training.py", line 378, in _make_predict_function
+    **kwargs)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/keras/engine/network.py", line 323, in __setattr__
+    super(Network, self).__setattr__(name, value)
+  File "/Users/kdr/python_project/flaskblog/venv/lib/python3.7/site-packages/keras/engine/base_layer.py", line 1215, in __setattr__
+    if not _DISABLE_TRACKING.value:
+AttributeError: '_thread._local' object has no attribute 'value'
+127.0.0.1 - - [20/Sep/2019 08:54:52] "GET /about HTTP/1.1" 200 -
